requirements.txt
# Quantum Computing Framework
qiskit==1.0.2
qiskit-machine-learning==0.7.2
qiskit-algorithms==0.3.0

# Deep Learning & Neural Networks
torch==2.2.0
torchvision==0.17.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Data Science & Numerical Computing
numpy==1.26.4
scikit-learn==1.4.1.post1

# Visualization & Monitoring
matplotlib==3.8.3
tqdm==4.66.2

# Utilities
pyyaml==6.0.1

